+++ nni/nas/execution/__init__.py
@@ -2,3 +2,4 @@
 # Licensed under the MIT license.
 
 from .api import *
+from .common import *
