+++ nni/nas/execution/common/engine.py
@@ -4,7 +4,7 @@
 from abc import ABC, abstractmethod, abstractclassmethod
 from typing import Any, Iterable, NewType, List, Union, Type
 
-from ..graph import Model, MetricData
+from .graph import Model, MetricData
 
 __all__ = [
     'GraphData', 'WorkerInfo',
