+++ nni/nas/execution/common/integration_api.py
@@ -1,6 +1,8 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
+__all__ = ['get_advisor', 'register_advisor', 'send_trial', 'receive_trial_parameters', 'get_experiment_id']
+
 import warnings
 from typing import NewType, Any
