+++ nni/nas/execution/common/utils.py
@@ -1,16 +1,21 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
+__all__ = ['get_mutation_dict', 'mutation_dict_to_summary', 'get_mutation_summary']
+
 from typing import Any, List
-from ..graph import Model
+from .graph import Model
+
 
 def _unpack_if_only_one(ele: List[Any]):
     if len(ele) == 1:
         return ele[0]
     return ele
 
+
 def get_mutation_dict(model: Model):
     return {mut.mutator.label: _unpack_if_only_one(mut.samples) for mut in model.history}
+
 
 def mutation_dict_to_summary(mutation: dict) -> dict:
     mutation_summary = {}
@@ -22,6 +27,7 @@ def mutation_dict_to_summary(mutation: dict) -> dict:
             for i, sample in enumerate(samples):
                 mutation_summary[f'{label}_{i}'] = sample
     return mutation_summary
+
 
 def get_mutation_summary(model: Model) -> dict:
     mutation = get_mutation_dict(model)
