+++ nni/nas/hub/pytorch/nasbench201.py
@@ -6,8 +6,8 @@ from typing import Callable, Dict
 import torch
 import torch.nn as nn
 
-from nni.retiarii import model_wrapper
-from nni.retiarii.nn.pytorch import NasBench201Cell
+from nni.nas import model_wrapper
+from .modules.nasbench201 import NasBench201Cell
 
 
 __all__ = ['NasBench201']
