+++ nni/nas/hub/pytorch/nasnet.py
@@ -18,11 +18,11 @@ except ImportError:
 
 import torch
 
-import nni.retiarii.nn.pytorch as nn
-from nni.retiarii import model_wrapper
+import nni.nas.nn.pytorch as nn
+from nni.nas import model_wrapper
 
-from nni.retiarii.oneshot.pytorch.supermodule.sampling import PathSamplingRepeat
-from nni.retiarii.oneshot.pytorch.supermodule.differentiable import DifferentiableMixedRepeat
+from nni.nas.oneshot.pytorch.supermodule.sampling import PathSamplingRepeat
+from nni.nas.oneshot.pytorch.supermodule.differentiable import DifferentiableMixedRepeat
 
 from .utils.fixed import FixedFactory
 from .utils.pretrained import load_pretrained_weight
