+++ nni/nas/hub/pytorch/proxylessnas.py
@@ -5,8 +5,8 @@ import math
 from typing import Optional, Callable, List, Tuple, Iterator, Union, cast, overload
 
 import torch
-import nni.retiarii.nn.pytorch as nn
-from nni.retiarii import model_wrapper
+import nni.nas.nn.pytorch as nn
+from nni.nas import model_wrapper
 
 from .utils.fixed import FixedFactory
 from .utils.pretrained import load_pretrained_weight
