+++ nni/nas/nn/pytorch/cell.py
@@ -12,8 +12,8 @@ except ImportError:
 import torch
 import torch.nn as nn
 
-from .api import ChosenInputs, LayerChoice, InputChoice
-from .nn import ModuleList  # pylint: disable=no-name-in-module
+from .choice import ChosenInputs, LayerChoice, InputChoice
+from .layers import ModuleList  # pylint: disable=no-name-in-module
 from .mutation_utils import generate_new_label
