+++ nni/nas/oneshot/pytorch/supermodule/differentiable.py
@@ -15,7 +15,7 @@ import torch.nn.functional as F
 
 from nni.common.hpo_utils import ParameterSpec
 from nni.nas.nn.pytorch import LayerChoice, InputChoice, ChoiceOf, Repeat
-from nni.nas.nn.pytorch.api import ValueChoiceX
+from nni.nas.nn.pytorch.choice import ValueChoiceX
 from nni.nas.nn.pytorch.cell import preprocess_cell_inputs
 
 from .base import BaseSuperNetModule
