+++ nni/nas/strategy/hpo.py
@@ -7,7 +7,7 @@ import logging
 import time
 from typing import Optional
 
-from .. import Sampler, submit_models, query_available_resources, is_stopped_exec, budget_exhausted
+from nni.nas import Sampler, submit_models, query_available_resources, is_stopped_exec, budget_exhausted
 from .base import BaseStrategy
 
 _logger = logging.getLogger(__name__)
